+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log (@egomobile/swagger-proxy)
 
+## 0.5.0
+
+- fix [setupSwaggerProxy() function](https://egomobile.github.io/node-swagger-proxy/modules.html#setupSwaggerProxy)
+- **BREAKING CHANGE:** remove `defaultPath` from [ISetupSwaggerProxyOptions interface](https://egomobile.github.io/node-swagger-proxy/interfaces/ISetupSwaggerProxyOptions.html)
+
 ## 0.4.0
 
 - add `defaultPath` to [ISetupSwaggerProxyOptions interface](https://egomobile.github.io/node-swagger-proxy/interfaces/ISetupSwaggerProxyOptions.html)